Create Matchsticks game .py
Matchsticks Game
This repository contains the solution to the Matchsticks Game problem. The Matchsticks Game is a two-player game where a group of matchsticks is placed on a table, and players take turns removing a certain number of matchsticks. The player who takes the last matchstick wins the game.

In this implementation, we focus on the scenario where Player A starts the game and we need to determine the optimal number of matchsticks A should pick on his first turn to guarantee a win, assuming both players play optimally.

Problem Description
Given the total number of matchsticks N, the task is to determine the number of matchsticks A should pick on his first turn to ensure his victory. If it is impossible for A to win, the function returns -1.

The rules of the game are as follows:

Each player can pick any number of matchsticks from 1 to 4 (both inclusive) during their turn.
The players take turns removing matchsticks.
The player who takes the last matchstick wins the game.
Solution Approach
The solution approach is based on the observation that A can always guarantee a win if the total number of matchsticks (N) is not divisible by 5. In this case, A can select a number of matchsticks such that the remaining count becomes non-divisible by 5, ensuring A's victory.

On the other hand, if N is divisible by 5, it is impossible for A to win since B can always select a number of matchsticks such that the remaining count becomes divisible by 5, giving B the opportunity to win.

The matchGame function implements this approach. It takes the total number of matchsticks (N) as input and returns the number of matchsticks A should pick on his first turn if it is possible for A to win. If it is impossible for A to win, the function returns -1.