Update Lemonade_Change.py
Lemonade Change
This repository contains the solution to the Lemonade Change problem. In the Lemonade Change problem, you, as the owner of a lemonade island, need to provide change to customers who are buying lemonade using $5, $10, or $20 bills. The goal is to determine if you can provide the correct change to every customer, where the net transaction should be that the customer pays $5.

Problem Description
You are given an integer array bills representing the bills that each customer pays in the order of their arrival. The bills can be either $5, $10, or $20. You need to determine if it is possible to provide every customer with the correct change. The following rules apply:

You start with no bills to provide change with.
Each customer buys only one lemonade.
To provide change, you can use the bills you received from previous customers.
Solution Approach
The solution to the Lemonade Change problem can be achieved by simulating the transaction process. We iterate through the bills array and keep track of the number of $5 and $10 bills we have. Whenever a customer pays with a $5 bill, we increment the count of $5 bills. When a customer pays with a $10 bill, we decrement the count of $5 bills and increment the count of $10 bills. If a customer pays with a $20 bill, we try to give change using a $10 bill and a $5 bill first (if we have both). If not possible, we give change using three $5 bills. If we cannot provide change in any of these ways, it means we do not have enough change, and we return False. If we successfully serve all the customers, we return True.

The lemonadeChange function implements this approach. It takes the total number of customers N and the array of bills bills as input and returns True if it is possible to provide the correct change to every customer, and False otherwise.