Create Pairwise swap elements of a linked list.py
We start by defining a class called Solution, which will contain methods for solving a specific problem related to linked lists.

Within the Solution class, we have a method named pairWiseSwap that takes the head of a singly linked list as its input.

We first check if the linked list is either empty (contains no nodes) or has only one node. If either condition is true, there are no pairs of nodes to swap, so we return the linked list as is.

We initialize three pointers: current, previous, and following.

current initially points to the head of the linked list.
previous is set to None initially and will be used to keep track of the previous node in the iteration.
following is also set to None initially and will be used to keep track of the node that follows current.
We enter a while loop that continues as long as both current and current.next are not None. This loop allows us to traverse the linked list, swapping pairs of nodes.

Inside the loop:

We set following to the node that comes after current.
We handle updating the head of the linked list or the next reference of the previous node:
If previous is None, it means we are at the beginning of the list, so we update the head to point to following.
Otherwise, if previous is not None, we update previous.next to point to following.
We update current.next to skip over the following node. This effectively detaches current from following and prepares current for the next pair of nodes to be swapped.

We update the next reference of following to point to current, effectively swapping the positions of current and following in the linked list.

We then update previous to be the previous current node, and we move current to the next node in the list to continue the pairwise swapping process.

The loop continues until we've processed all pairs of nodes in the linked list.

Finally, the function returns the updated head of the linked list, which is now the new head after pairwise swapping has been applied.